VFB-443: refactored reports to use signpostingCallOptions from parcels
